refactor(react-sdk): organize token hooks into domain folders
Move hooks from flat token/ into categorized folders:
- authorization/ (allow, revoke, session)
- balance/ (confidential balance, activity feed, optimistic updates)
- transfer/ (confidential transfer, approve)
- shield/ (shield, approve underlying, allowance)
- unshield/ (unshield, resume)
- unwrap/ (unwrap, finalize)
- delegation/ (delegate decryption, revoke delegation, status)
- token/ (metadata, discovery, fees, supply, core token hooks)
