+++ packages/react-sdk/src/__tests__/provider-hooks-extended.test.tsx
@@ -1,9 +1,9 @@
 import { describe, expect, it, vi } from "../test-fixtures";
 import { waitFor } from "@testing-library/react";
 import type { Address } from "@zama-fhe/sdk";
-import { useUnderlyingAllowance } from "../token/use-underlying-allowance";
-import { useUnshield } from "../token/use-unshield";
-import { useUnshieldAll } from "../token/use-unshield-all";
+import { useUnderlyingAllowance } from "../shield/use-underlying-allowance";
+import { useUnshield } from "../unshield/use-unshield";
+import { useUnshieldAll } from "../unshield/use-unshield-all";
 import { useMetadataSuspense } from "../token/use-metadata";
 import { useTotalSupplySuspense } from "../token/use-total-supply";
 import { useWrapperDiscoverySuspense } from "../token/use-wrapper-discovery";
