+++ packages/react-sdk/src/shield/use-underlying-allowance.ts
@@ -4,7 +4,7 @@ import { useQuery, useSuspenseQuery } from "../utils/query";
 import type { UseQueryOptions } from "@tanstack/react-query";
 import type { Address } from "@zama-fhe/sdk";
 import { signerAddressQueryOptions, underlyingAllowanceQueryOptions } from "@zama-fhe/sdk/query";
-import { useReadonlyToken } from "./use-readonly-token";
+import { useReadonlyToken } from "../token/use-readonly-token";
 
 export { underlyingAllowanceQueryOptions };
