+++ packages/react-sdk/src/unshield/use-resume-unshield.ts
@@ -5,18 +5,17 @@ import { useMutation } from "@tanstack/react-query";
 import type { Address, TransactionResult } from "@zama-fhe/sdk";
 import {
   invalidateAfterUnshield,
-  type UnshieldParams,
-  unshieldMutationOptions,
+  type ResumeUnshieldParams,
+  resumeUnshieldMutationOptions,
 } from "@zama-fhe/sdk/query";
-import { useToken, type UseZamaConfig } from "./use-token";
+import { useToken, type UseZamaConfig } from "../token/use-token";
 
 /**
- * Unshield a specific amount and finalize in one call.
- * Orchestrates: unwrap → wait for receipt → parse event → finalize.
+ * Resume an interrupted unshield from an existing unwrap tx hash.
+ * Useful when the user submitted the unwrap but the finalize step was
+ * interrupted (e.g. page reload, network error).
  *
  * Errors are {@link ZamaError} subclasses — use `instanceof` to handle specific failures:
- * - {@link SigningRejectedError} — user rejected the wallet prompt
- * - {@link EncryptionFailedError} — FHE encryption failed during unwrap
  * - {@link DecryptionFailedError} — public decryption failed during finalize
  * - {@link TransactionRevertedError} — on-chain transaction reverted
  *
@@ -25,18 +24,18 @@ import { useToken, type UseZamaConfig } from "./use-token";
  *
  * @example
  * ```tsx
- * const unshield = useUnshield({ tokenAddress: "0x...", wrapperAddress: "0x..." });
- * unshield.mutate({ amount: 500n });
+ * const resumeUnshield = useResumeUnshield({ tokenAddress: "0x...", wrapperAddress: "0x..." });
+ * resumeUnshield.mutate({ unwrapTxHash: "0xabc..." });
  * ```
  */
-export function useUnshield(
+export function useResumeUnshield(
   config: UseZamaConfig,
-  options?: UseMutationOptions<TransactionResult, Error, UnshieldParams, Address>,
+  options?: UseMutationOptions<TransactionResult, Error, ResumeUnshieldParams, Address>,
 ) {
   const token = useToken(config);
 
-  return useMutation<TransactionResult, Error, UnshieldParams, Address>({
-    ...unshieldMutationOptions(token),
+  return useMutation<TransactionResult, Error, ResumeUnshieldParams, Address>({
+    ...resumeUnshieldMutationOptions(token),
     ...options,
     onSuccess: (data, variables, onMutateResult, context) => {
       options?.onSuccess?.(data, variables, onMutateResult, context);
