+++ packages/react-sdk/src/unwrap/use-unwrap.ts
@@ -4,17 +4,20 @@ import type { UseMutationOptions } from "@tanstack/react-query";
 import { useMutation } from "@tanstack/react-query";
 import type { Address, TransactionResult } from "@zama-fhe/sdk";
 import {
-  confidentialApproveMutationOptions,
-  invalidateAfterApprove,
-  type ConfidentialApproveParams,
+  invalidateAfterUnwrap,
+  type UnwrapParams,
+  unwrapMutationOptions,
 } from "@zama-fhe/sdk/query";
-import { useToken, type UseZamaConfig } from "./use-token";
+import { useToken, type UseZamaConfig } from "../token/use-token";
 
 /**
- * Set operator approval for a confidential token. Defaults to 1 hour.
+ * Request an unwrap for a specific amount. Encrypts the amount first.
+ * Call {@link useFinalizeUnwrap} after the request is processed on-chain,
+ * or use {@link useUnshield} for a single-call orchestration.
  *
  * Errors are {@link ZamaError} subclasses — use `instanceof` to handle specific failures:
  * - {@link SigningRejectedError} — user rejected the wallet prompt
+ * - {@link EncryptionFailedError} — FHE encryption of the unwrap amount failed
  * - {@link TransactionRevertedError} — on-chain transaction reverted
  *
  * @param config - Token address (and optional wrapper) identifying the token.
@@ -22,22 +25,22 @@ import { useToken, type UseZamaConfig } from "./use-token";
  *
  * @example
  * ```tsx
- * const approve = useConfidentialApprove({ tokenAddress: "0x..." });
- * approve.mutate({ spender: "0xOperator" });
+ * const unwrap = useUnwrap({ tokenAddress: "0x..." });
+ * unwrap.mutate({ amount: 500n });
  * ```
  */
-export function useConfidentialApprove(
+export function useUnwrap(
   config: UseZamaConfig,
-  options?: UseMutationOptions<TransactionResult, Error, ConfidentialApproveParams, Address>,
+  options?: UseMutationOptions<TransactionResult, Error, UnwrapParams, Address>,
 ) {
   const token = useToken(config);
 
-  return useMutation<TransactionResult, Error, ConfidentialApproveParams, Address>({
-    ...confidentialApproveMutationOptions(token),
+  return useMutation<TransactionResult, Error, UnwrapParams, Address>({
+    ...unwrapMutationOptions(token),
     ...options,
     onSuccess: (data, variables, onMutateResult, context) => {
       options?.onSuccess?.(data, variables, onMutateResult, context);
-      invalidateAfterApprove(context.client, config.tokenAddress);
+      invalidateAfterUnwrap(context.client, config.tokenAddress);
     },
   });
 }
